SlashBreak: Fix long instead of int being passed through to modal

+++ src/main/java/dev/skullition/lockium/command/GtCommands.java
@@ -440,7 +440,7 @@ public class GtCommands {
                         .setValue("0")
                         .setRequiredRange(1, 1)
                         .build()))
-            .bindTo(SlashBreakModal.MODAL_NAME, itemDetail, itemQuery, blockCount)
+            .bindTo(SlashBreakModal.MODAL_NAME, itemDetail, itemQuery, (int) blockCount)
             .build();
     event.replyModal(modal).queue();
   }
